requirements.txt
flask
gunicorn
sklearn
numpy
pandas
PyYAML
-e .